requirements.txt
streamlit
langchain
langchain-community
langchain-openai
pypdf
faiss-cpu
python-dotenv
